requirements.txt
# Autonomous Trading & Portfolio Optimization Agent
# Bachelor Thesis - Multi-Agent System

# Core orchestration
crewai>=0.80.0
crewai-tools>=0.14.0

# LLM (Groq via LiteLLM — CrewAI routes groq/* through LiteLLM)
groq>=0.4.0
litellm>=1.50.0

# Market data & analysis
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.24.0

# Frontend
streamlit>=1.28.0
plotly>=5.0.0

# PDF report generation
fpdf2>=2.7.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
PyYAML>=6.0
